split ebpf to interface and implementation
