+++ processmanager/ebpf/impl/ebpf_integration_test.go
@@ -3,7 +3,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package ebpf
+package impl
 
 import (
 	"testing"
